projects/RSG/requirements.txt
gunicorn
pandas
PyPDF2
requests
config
google.cloud.storage
google.cloud.bigquery
Flask
pandas_gbq
openpyxl
pyxlsb
xlrd
regex
numpy
img2pdf
google-cloud-documentai
google-cloud
googleapis-common-protos
